requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.0
azure-core==1.29.3
azure-identity==1.14.0
azure-storage-blob==12.17.0
billiard==3.6.4.0
black==23.7.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
dnspython==2.2.1
docxcompose==1.3.6
docxtpl==0.16.4
drf-base64==2.0
email-validator==1.3.0
environ==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
msal==1.23.0
msal-extensions==1.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
OdooRPC==0.9.0
openpyxl==3.0.10
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.9.1
pluggy==1.0.0
portalocker==2.7.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.8.0
PyPDF2==3.0.1
python-docx==0.8.11
pytz==2022.6
redis==4.4.0
requests==2.29.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.5
wrapt==1.15.0
